# openapi.yaml
openapi: 3.1.0
info:
  title: Blue Economy Port Interoperability API
  version: 0.1.0
  description: >-
    Implemented local port-call workflow; IMO/NSW partner profiles remain
    separate acceptance gates. The eCallUp booking, slot, gate-scan, call-up
    queue and NSW ingress routes (/v1/bookings, /v1/slots, /v1/terminals,
    /v1/queue-requests, /v1/terminals/{id}/queue, /v1/gate/scans,
    /v1/nsw/port-calls) are wired and tenant-isolated but are documented in
    README.md until the Ministry approves their external interface profile.
    All /v1 routes require the loopback trusted-proxy headers plus a validated
    HS256 gateway tenant token (Authorization: Bearer); the NSW ingress route
    instead requires an RS256 authority JWS in X-NSW-Signature.
servers:
  - url: http://localhost:18080
paths:
  /healthz:
    get:
      operationId: health
      responses:
        '200':
          description: Service is alive.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    const: ok
  /v1/port-calls:
    post:
      operationId: createPortCall
      security: [loopbackTrustedProxy: []]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TrustedProxy'
        - $ref: '#/components/parameters/AuthenticatedPrincipal'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortCallRequest'
      responses:
        '201':
          description: Created or exact idempotent replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortCall'
        '400': {description: Invalid request or missing idempotency key.}
        '401': {description: Missing verified local edge identity.}
        '409': {description: Conflicting idempotency reuse.}
  /v1/port-calls/{call_id}:
    get:
      operationId: getPortCall
      security: [loopbackTrustedProxy: []]
      parameters:
        - $ref: '#/components/parameters/CallID'
        - $ref: '#/components/parameters/TrustedProxy'
        - $ref: '#/components/parameters/AuthenticatedPrincipal'
      responses:
        '200':
          description: Port call.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PortCall'}
        '401': {description: Missing verified local edge identity.}
        '404': {description: Port call not found.}
  /v1/port-calls/{call_id}/submit:
    post:
      operationId: submitPortCall
      security: [loopbackTrustedProxy: []]
      parameters:
        - $ref: '#/components/parameters/CallID'
        - $ref: '#/components/parameters/TrustedProxy'
        - $ref: '#/components/parameters/AuthenticatedPrincipal'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TransitionRequest'}
      responses:
        '200': {description: Port call submitted.}
        '409': {description: Version or state conflict.}
  /v1/port-calls/{call_id}/accept:
    post:
      operationId: acceptPortCall
      security: [loopbackTrustedProxy: []]
      parameters:
        - $ref: '#/components/parameters/CallID'
        - $ref: '#/components/parameters/TrustedProxy'
        - $ref: '#/components/parameters/AuthenticatedPrincipal'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TransitionRequest'}
      responses:
        '200': {description: Port call accepted.}
        '409': {description: Version or state conflict.}
  /v1/port-calls/{call_id}/reject:
    post:
      operationId: rejectPortCall
      security: [loopbackTrustedProxy: []]
      parameters:
        - $ref: '#/components/parameters/CallID'
        - $ref: '#/components/parameters/TrustedProxy'
        - $ref: '#/components/parameters/AuthenticatedPrincipal'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TransitionRequest'}
      responses:
        '200': {description: Port call rejected.}
        '409': {description: Version or state conflict.}
  /v1/partner-capabilities:
    get:
      operationId: getPartnerCapabilities
      security: [loopbackTrustedProxy: []]
      parameters:
        - $ref: '#/components/parameters/TrustedProxy'
        - $ref: '#/components/parameters/AuthenticatedPrincipal'
      responses:
        '200': {description: Source-backed local capabilities.}
        '401': {description: Missing verified edge identity.}
  /v1/port-calls/{call_id}/documents:
    post:
      operationId: declareDocument
      security: [loopbackTrustedProxy: []]
      parameters: [{$ref: '#/components/parameters/CallID'}, {$ref: '#/components/parameters/TrustedProxy'}, {$ref: '#/components/parameters/AuthenticatedPrincipal'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DocumentDeclarationRequest'}}}}
      responses: {'201': {description: Declared document.}, '409': {description: Replay/conflict.}, '422': {description: Invalid workflow.}}
  /v1/port-calls/{call_id}/documents/review:
    post:
      operationId: reviewDocument
      security: [loopbackTrustedProxy: []]
      parameters: [{$ref: '#/components/parameters/CallID'}, {$ref: '#/components/parameters/TrustedProxy'}, {$ref: '#/components/parameters/AuthenticatedPrincipal'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DocumentReviewRequest'}}}}
      responses: {'200': {description: Reviewed document.}, '409': {description: Version/state conflict.}}
  /v1/port-calls/{call_id}/documents/supersede:
    post:
      operationId: supersedeDocument
      security: [loopbackTrustedProxy: []]
      parameters: [{$ref: '#/components/parameters/CallID'}, {$ref: '#/components/parameters/TrustedProxy'}, {$ref: '#/components/parameters/AuthenticatedPrincipal'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DocumentSupersessionRequest'}}}}
      responses: {'201': {description: Immutable supersession evidence recorded.}, '409': {description: Document conflict.}}
  /v1/port-calls/{call_id}/clearance:
    post:
      operationId: decideClearance
      security: [loopbackTrustedProxy: []]
      parameters: [{$ref: '#/components/parameters/CallID'}, {$ref: '#/components/parameters/TrustedProxy'}, {$ref: '#/components/parameters/AuthenticatedPrincipal'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ClearanceDecisionRequest'}}}}
      responses: {'201': {description: Clearance decision.}, '409': {description: Version/decision conflict.}, '422': {description: Evidence prerequisite unmet.}}
  /v1/port-calls/{call_id}/clearance/amend:
    post:
      operationId: amendClearance
      security: [loopbackTrustedProxy: []]
      parameters: [{$ref: '#/components/parameters/CallID'}, {$ref: '#/components/parameters/TrustedProxy'}, {$ref: '#/components/parameters/AuthenticatedPrincipal'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ClearanceAmendmentRequest'}}}}
      responses: {'200': {description: Amended clearance.}, '409': {description: Version/actor/decision conflict.}, '422': {description: Invalid amendment.}}
components:
  securitySchemes:
    loopbackTrustedProxy:
      type: apiKey
      in: header
      name: X-Authenticated-Principal
      description: Local-only boundary; production requires a verified Ministry OIDC/API-edge profile.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 256}
    TrustedProxy:
      name: X-Trusted-Proxy
      in: header
      required: true
      schema: {type: string, const: loopback}
    AuthenticatedPrincipal:
      name: X-Authenticated-Principal
      in: header
      required: true
      schema: {type: string, minLength: 1}
    CallID:
      name: call_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 256}
  schemas:
    CreatePortCallRequest:
      type: object
      additionalProperties: false
      required: [call_id, vessel_imo, port_code, declaration_reference, submitted_by]
      properties:
        call_id: {type: string, minLength: 1, maxLength: 256}
        vessel_imo: {type: string, pattern: '^[0-9]{7}$'}
        port_code: {type: string, pattern: '^[A-Z]{2,8}$'}
        declaration_reference: {type: string, minLength: 1, maxLength: 256}
        submitted_by: {type: string, minLength: 1, maxLength: 256}
    TransitionRequest:
      type: object
      additionalProperties: false
      required: [expected_version]
      properties:
        expected_version: {type: integer, minimum: 1}
    DocumentDeclarationRequest:
      type: object
      additionalProperties: false
      required: [document_type, media_type, size_bytes, sha256, declared_by]
      properties:
        document_type: {type: string, pattern: '^[A-Za-z][A-Za-z0-9._:-]{0,63}$'}
        media_type: {type: string, pattern: '^[A-Za-z0-9.+-]+/[A-Za-z0-9.+-]+$'}
        size_bytes: {type: integer, minimum: 1, maximum: 104857600}
        sha256: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        declared_by: {type: string, minLength: 1, maxLength: 256}
    DocumentReviewRequest:
      type: object
      additionalProperties: false
      required: [document_id, expected_version, status, reviewed_by, reason]
      properties:
        document_id: {type: string, format: uuid}
        expected_version: {type: integer, minimum: 1}
        status: {type: string, enum: [VERIFIED, REJECTED]}
        reviewed_by: {type: string, minLength: 1, maxLength: 256}
        reason: {type: string, minLength: 1, maxLength: 1024}
    DocumentSupersessionRequest:
      type: object
      additionalProperties: false
      required: [original_document_id, replacement_document_id, reason, superseded_by]
      properties:
        original_document_id: {type: string, format: uuid}
        replacement_document_id: {type: string, format: uuid}
        reason: {type: string, minLength: 1, maxLength: 1024}
        superseded_by: {type: string, minLength: 1, maxLength: 256}
    ClearanceDecisionRequest:
      type: object
      additionalProperties: false
      required: [expected_version, decision, reason, decided_by]
      properties:
        expected_version: {type: integer, minimum: 1}
        decision: {type: string, enum: [APPROVED, REJECTED]}
        reason: {type: string, minLength: 1, maxLength: 1024}
        decided_by: {type: string, minLength: 1, maxLength: 256}
    ClearanceAmendmentRequest:
      type: object
      additionalProperties: false
      required: [expected_version, decision, reason, amended_by]
      properties:
        expected_version: {type: integer, minimum: 1}
        decision: {type: string, enum: [APPROVED, REJECTED]}
        reason: {type: string, minLength: 1, maxLength: 1024}
        amended_by: {type: string, minLength: 1, maxLength: 256}
    PortCall:
      type: object
      required: [call_id, vessel_imo, port_code, declaration_reference, submitted_by, status, created_at, updated_at, version]
      properties:
        call_id: {type: string}
        vessel_imo: {type: string}
        port_code: {type: string}
        declaration_reference: {type: string}
        submitted_by: {type: string}
        status: {type: string, enum: [DRAFT, SUBMITTED, ACCEPTED, REJECTED]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        version: {type: integer, minimum: 1}